# LoggingExecutionContextScopeBuilderAnalyzer optimizations