With ImmutabilityAnalyzer as reference (given run variability):

| Analyzer | Before | After |
| -- | -- | -- |
| ImmutabilityAnalyzer | 11.804s | 11.681s |
| LoggingExecutionContextScopeBuilderAnalyzer | 3.136s | 194ms |